app/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
databases==0.8.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
pathlib==1.0.1
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pydentic==0.0.1.dev3
pytest==7.4.4
python-dotenv==1.0.0
python-stdnum==1.19
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
pydantic[email]
python-multipart
casbin
casbin_sqlalchemy_adapter
fastapi-mail
APScheduler==3.10.4
itsdangerous
python-jose
passlib
databases[postgresql]
faker
factory-boy